i18n + some path refactored + some bugs ;)
