+++ config/locales/ru.yml
@@ -1,0 +1,4 @@
+ru:
+  pandora:
+    some:
+      text: "foobar"
