+++ lib/pandora.rb
@@ -46,6 +46,10 @@ module Pandora
     @@root_dir ||= File.expand_path("../../", __FILE__)
   end
 
+  def self.t(key)
+    ::I18n.t('pandora.' + key)
+  end
+
   # Pandora's directories
   %w(base view model lang util files).each do |method_name|
     class_eval "def self.#{method_name}_dir; File.join root, '#{method_name}'; end"
